v4

+++ algo_rec/rank/prod_v4/feature_serv_describe_tfv2.py
@@ -23,7 +23,7 @@ feature_describe_pred = {
     "mt_i2i_main_score": v1.FixedLenFeature(1, tf.float32, -1.0),
     "mt_i2i_long_score": v1.FixedLenFeature(1, tf.float32, -1.0),
     "mt_i2i_short_score": v1.FixedLenFeature(1, tf.float32, -1.0),
-    "mt_u2i_f": v1.FixedLenFeature(1, tf.float32, 0),
+    "mt_u2i_f": v1.FixedLenFeature(1, tf.int64, 0),
     "mt": v1.FixedLenFeature(6, tf.string, default_value=[""] * 6),
 
     "main_goods_id": v1.FixedLenFeature(1, tf.string, "-1"),
